fixed routing issues and project setup
