+++ src/App.css
@@ -5,46 +5,4 @@
   margin: 0;
   padding: 0;
   font-family: 'Montserrat', sans-serif;
-  background-color: #1a1a1c;
-  color: #fff;
-}
-
-.coin-app {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-  color: #fff
-}
-
-.coin-search {
-  margin-bottom: 64px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.coin-text {
-  margin-bottom: 32px;
-  text-align: center;
-}
-
-.coin-input {
-  padding-left: 16px;
-  width: 300px;
-  height: 50px;
-  border-radius: 4px;
-  border: none;
-  background-image: linear-gradient(
-    -225deg,
-    #ac32e4 0%,
-    #7918f2 48%,
-    #4801ff 100%
-  );
-  color: #fff
-}
-
-.coin-input::placeholder {
-  color: #fff;
 }
