+++ src/pages/index.js
@@ -1,16 +1,51 @@
-import React from 'react'
+import React, {useState, useEffect} from 'react';
+import axios from 'axios';
+import Coin from './coin/Coin';
+import './Home.css';
 
-const Home = () => {
+function Home() {
+    const [coins, setCoins] = useState([])
+    const [search, setSearch] = useState('')
+
+    useEffect(() => {
+        axios.get('https://api.coingecko.com/api/v3/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=100&page=1&sparkline=false')
+        .then(res => {
+        setCoins(res.data);
+        }).catch(error => console.log(error))
+    }, []);
+
+    const handleChange = e => {
+        setSearch(e.target.value)
+    }
+
+    const filteredCoins = coins.filter(coin =>
+        coin.name.toLowerCase().includes(search.toLowerCase())
+        )
+
     return (
-        <div 
-            style={{display: 'flex', 
-            justifyContent: 'center', 
-            alignItems: 'center', 
-            height: '90vh'
-        }}
-        >
-            <h1>Home</h1>
-        </div>
+    <div className='coin-app'>
+      <div className='coin-search'>
+        <h1 className='coin-text'>Crypto Tracker</h1>
+        <form>
+          <input type='text' placeholder='Search'
+          className='coin-input' onChange=
+          {handleChange}/>
+        </form>
+      </div>
+      {filteredCoins.map(coin => {
+        return <Coin
+          rank={coin.market_cap_rank}
+          key={coin.id} 
+          name={coin.name} 
+          image={coin.image}
+          symbol={coin.syymbol}
+          marketcap={coin.market_cap}
+          price={coin.current_price}
+          priceChange={coin.price_change_percentage_24h}
+          volume={coin.total_volume}
+          />;
+      })}
+    </div>
     )
 }
